Unit and functional test for transaction create
This is not really a meaningful test yet, but it illustrates the
difference between the unit tests and the functional tests.

The unit tests really only test a single function by giving an input and
checking the return value.

The functional tests are integration tests. These create mock request
to the actual WSGI app that is created. The response needs to be tested
like what is received by the browser. So this also tests what is done
with the response in the template.

It took me forever to figure out how to create the app instance. It was
always failing because it was missing the ``sqlalchemy.url`` setting.
The solution is that the ``settings dictionary`` that is passed into the
``main()`` call needs to be unpacked: ::

	testapp = main({}, **settings)