Remove functional test for list view
For some reason, I can not get the database to initialize on the
functional test. I assume it has something to do with the initialization
of the app, but I do not know. That is the only difference between the
unittests and the functional tests.

The functional tests would be nice, because they would allow to test
the final page (the rendered template) like it would show up in the
browser, but without the need to actually launch the browser.

I guess I can forgo this type of test for now and only test the view
functions directly.

It would be nice to use the functional tests too (becuase the ensure
the functionality for the user), but I have already wasted way to much
time on that.