Raise HTTPNotFound, not return
So the "exception response handler" is only invoked when the
HTTP Exception is raised, not when it is returned.

I guess that makes more sense. The exception is an unexpected event,
kinda. It is not the normal path that the function want to take.
Since Pyramid defines all these HTTP response types as exceptions,
it is probably most consistent to raise the corresponding exception,
instead of returning it.

This now makes use of the custom 404 view when the transaction for
which the detail is requested does not exist.