Attach engine to registry to retrieve in testsetup
Before, I could not find a way to create the SQLite in-memory database
during functional testing. Even calling `Base.metadata.create_all()`
did not work.

The issue was, that I was creating an engine in the test setup code.
This engine was not the same one that the app uses internally. Therefore
I would have two in-memory databases. One that was initialized during
the testing setup and one that was used by the app code.

The one used by the app was never initialized though. For the normal
SQLite development database, that initialization is done with `alembic`.

To make use of the same in-memory database in the testing setup and in
the app code, I need a way to retrieve the engine used by the app.
To achieve this, I added the engine used by the app to the app registry.

During test setup, I am instantiating the app and thus have access to
the registry like so:
```python
from heath import main
app = main({}, **settings)
engine = app.registry["engine"]
```

This allows me to initialize the database with
`Base.metadata.create_all(engine)` in the test setup code.