Amount float column
Alembic can for some reason not deal well with dropping SQLite columns.
I wanted to change the type of the amount column from numeric to float,
because numeric leads to sqlalchemy warning about rounding etc.

To get the desired change, I deleted all migrations and starting fresh
with the correct configuration. Otherwise this would just be a pain in
my ass.