Move database initialization to BaseTest setup
Before, every test that relied on a working database had to specifically
initialize it. This was tedious, because that was almost every test.

Now the database is automatically initialized in the BaseTests setUp
method. Thus, the initialization is removed from the other test methods.

One of the example tests relied on a missing database connection.
Because the examples are not needed anymore, they are also removed.
This also makes the tests module structure cleaner.