Redirect after create raised not returned
Before, the create view returned the HTTPFound exception to redirect the
user. I think this is inconsistent with a view that usually returns
data.

To make the behavior more consistent, the HTTPFound is now raised and
no data is returned.

Tests usually expect a test to be failed if an exception is raised.
To prevent the test failure, the exception has to be expected.
This is done by wrapping the view call with within a context manager:
```python
with self.assertRaises(HTTPFound):
    create(request)
```