Return HTTPFound after delete
The functional test revealed that then HTTPFound is raised, that the
desired change is no correctly persisted in the database.

This issue is caused by the transaction manager, which is attached to
the request. The transaction manager aborts ongoing transactions when
an exception is raised. To prevent the abort, and have the changes
persist in the database, the redirect after a successful delete needs
to be returned rather than raised.

Similar changes will be necessary in other view functions that currently
raise HTTPFound after successful actions.