Add test for delete form functionality
The previously created tests used explicit post requests to the delete
endpoint to test the functionality of the view.

The newly added test only makes use of the rendered elements (in this
case only the submit button). This test is even closer to the user
experience, but easier than implementing tests with browser driver such
as selenium.