Test links in list endpoint with BeautifulSoup
Instead of just looking for the links in the whole body, with
BeautifulSoup it is easy to first reduce the search string to only the
link urls.

This makes the test a little more concise and descriptive. Otherwise,
the test might pass because some thing that looks like the link is
somewhere in the body.