Refactoring test layout
Functional and unit tests have very different scopes and requirements.

Functional tests should be black box tests. They are not concerned with
implementation details and the like.The only test functionality, mainly
from a user's perspective. How the functionality is reached is not
important.

This case is different for unit tests. Unit tests are white box tests.
They can have and probably require some knowledge of the inner workings
of the unit (function) under test. Even to mock other resources and
functions during a unit tests, I need to know about the resource being
used in the test.

This is why it makes sense that the structure of the unit tests maps the
structure of the code closely. That makes no sense on the other hand for
the functional tests. These tests are not concerned with the inner
workings or even the code structure. Therefore, functional tests should
be organized by function (from a user's perspective) and not by code
structure.

This transition in the testing structure is started with this commit.