Mock route_url on DummyRequest
View functions often result in redirects, especially the ones that are
handling post requests. Once the creation or update operation has
succeeded, you want to direct the user to a different page.

It is a good practice to generate the URLs for the redirect from the
route config, rather than hard coding the URLs in the view functions.
This is, because many view functions might redirect to the same URL.
If that URL should ever change, all these view functions have to be
updated.

Unfortunately, the URL generation from route names does not working in
unit tests of the view functions. This is, because there is no running
app in the unit tests. These really only test the return of a function
based on some input. Without the running app, there is no configuration
or place where the route names can be mapped to URLs.

Therefore, the URL lookup function fails in the unit tests, but works
fine in functional tests or when actually running the application.

Still, it makes sense to test the view functions more isolated, to not
be influenced by issues that might be caused by the template etc.

In the unit tests it usually does not matter where the redirect goes to.
The redirect is only a sign of a successful handling of the post
request. Therefore it is possible to "mock" the function that would
usually look up the URLs (based on route names).