Add account model
To group transactions and represent actual financial situations better,
an account model has been created.

To enable the creation of foreign key relationships, the tables need
to be altered. This is a problem for SQLite, because is only supports
minimal alterations of tables. To allow these kinds of alterations,
the migration tool needs to copy and move the data to a new table with
the desired structure. The new table is filled with the original data
and then renamed to the original value. To allow this, "batch"
migrations need to be enabled in the alembic configuration.
See also: https://alembic.sqlalchemy.org/en/latest/batch.html